% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_row_sums.R
\name{add_row_sums}
\alias{add_row_sums}
\title{Add Row-wise Sums to a Dataframe}
\usage{
add_row_sums(data, var_name, item_names, ignore_na = FALSE)
}
\arguments{
\item{data}{A dataframe or tibble.}

\item{var_name}{The name of the new variable to be created (as a string).}

\item{item_names}{A character vector of the column names to sum up.}

\item{ignore_na}{Logical; defaults to FALSE. If FALSE, any NA in the row results
in an NA sum. If TRUE, the sum ignores NAs.}
}
\value{
A new dataframe that includes the new variable and is ungrouped.
}
\description{
Add Row-wise Sums to a Dataframe
}
\examples{
# Example usage:
my_data <- data.frame(a = c(1, 2, NA), b = c(3, 4, 5))

my_data \%>\%
  add_row_sums(
    var_name = "total",
    item_names = c("a", "b"),
    ignore_na = TRUE
  )
}
