% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_df.R
\name{convert.df}
\alias{convert.df}
\title{Convert a SPSS Data Frame (imported with haven::) Into an Analysis-Ready One}
\usage{
convert.df(data)
}
\arguments{
\item{data}{Your data file that either has numeric or factor columns. The character columns will be NA.}
}
\value{
A data frame that is haven-labelled variables are converted.
}
\description{
Convert a SPSS Data Frame (imported with haven::) Into an Analysis-Ready One
}
\examples{
converted_dataframe <- convert.df(spss_imported_dataframe)
}
