% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_results.R
\name{get_model_results}
\alias{get_model_results}
\title{Get Summaries for the LM(ER) Models}
\usage{
get_model_results(models_list, the_model = "M_LMER_complex")
}
\arguments{
\item{models_list}{The model list derived from the RadLab::run_models() function.}

\item{the_model}{The best fitted model.}
}
\value{
Prints the summary() and anova() for each action option.
}
\description{
Get Summaries for the LM(ER) Models
}
