% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_means.R
\name{graph_means}
\alias{graph_means}
\title{Graph Vignette Means from Data}
\usage{
graph_means(data, name_data, n_vignettes)
}
\arguments{
\item{data}{a data frame with only data from vignette responses}

\item{name_data}{a data frame with names of vignettes and their corresponding order
(e.g., for row 1, column 1, it is "Bakery", and row 1, column 2 is 1)}

\item{n_vignettes}{Total number of vignettes}
}
\description{
Graph Vignette Means from Data
}
