% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_with_pilot.R
\name{merge_with_pilot}
\alias{merge_with_pilot}
\title{Merge Study with Pilot Scores}
\usage{
merge_with_pilot(excel_file, v_means_pilot, study_number)
}
\arguments{
\item{excel_file}{The Excel file that includes the order of the vignettes in each study.}

\item{v_means_pilot}{A data frame that includes the pilot study vignette means
(the extremity scores).}

\item{study_number}{The corresponding column number in the excel sheet.
Note that the first column in the excel sheet is the names of the vignettes.}
}
\value{
a data frame that matches the vignette names & order within the particular study
with the extremity scores (means) derived from the pilot study.
}
\description{
Merge Study with Pilot Scores
}
\examples{
o_study1 <- RadLab::merge_with_pilot(studies_all, v.means_pilot, 1)
}
