% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vignette_means.R
\name{plot_vignette_means}
\alias{plot_vignette_means}
\title{Graph Vignette Means from Data}
\usage{
plot_vignette_means(data, name_data, n_vignettes, label_x = TRUE)
}
\arguments{
\item{data}{A data frame with only data from vignette responses.}

\item{name_data}{A data frame with names of vignettes and their corresponding order
(e.g., for row 1, column 1, it is "Bakery", and row 1, column 2 is 1).}

\item{n_vignettes}{Total number of vignettes.}

\item{label_x}{Logical; if TRUE, labels the x-axis with vignette names. Defaults to TRUE.}
}
\value{
A ggplot2 object displaying the means of vignette responses across action types.
}
\description{
Graph Vignette Means from Data
}
\examples{
# Example usage:
plot_vignette_means(
  data = d_study1[, 1:60],
  name_data = o_study1,
  n_vignettes = 15
)
}
