% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_sums.R
\name{row_sums}
\alias{row_sums}
\title{Sum the Rows of a Dataframe}
\usage{
row_sums(data, var_name, item_names, ignore_na = F)
}
\arguments{
\item{data}{a dataframe}

\item{var_name}{the name of the new variable that is to be created}

\item{item_names}{the name of the items to sum up the points for}

\item{ignore_na}{FALSE by default:
if there is one NA, the total sum will be NA.
if TRUE, the sum ignores the NA and the sum only involves the non-NA.}
}
\value{
a new dataframe that includes the new variable
}
\description{
Sum the Rows of a Dataframe
}
