% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wide_to_long.R
\name{wide_to_long}
\alias{wide_to_long}
\title{Wide to Long Transformation for RadLab Vignette Data}
\usage{
wide_to_long(data, name_data, n_vignettes, IDvar = "ID")
}
\arguments{
\item{data}{the data frame that you want to transform from the wide format to the long format}

\item{name_data}{a data frame with names of vignettes and their corresponding order}

\item{n_vignettes}{Total number of vignettes}

\item{IDvar}{ID variable}
}
\value{
A long data frame suitable for modelling
}
\description{
Wide to Long Transformation for RadLab Vignette Data
}
\examples{
d_hartog_long <- d_hartog \%>\% RadLab::wide_to_long(
    name_data = sv_hartog,
    n_vignettes = 15)
}
