% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wide_to_long.R
\name{wide_to_long}
\alias{wide_to_long}
\title{Wide to Long Transformation for RadLab Vignette Data}
\usage{
wide_to_long(data, name_data, n_vignettes, IDvar = "ID")
}
\arguments{
\item{data}{A data frame where the first (n_vignettes * 4) columns are
interleaved vignette responses (ina, na, nna, enna).}

\item{name_data}{A data frame with names of vignettes and their
corresponding order (must contain 'order' and 'Vignette' columns).}

\item{n_vignettes}{Total number of vignettes (integer).}

\item{IDvar}{Character string specifying the name of the ID variable.
Defaults to "ID".}
}
\value{
A long-format data frame with columns: ID, time, ina, na, nna, enna,
and Vignette name.
}
\description{
Transforms interleaved vignette data from wide to long format.
**Note:** This function assumes that the vignette response columns
(4 per vignette) are the FIRST columns in the data frame.
}
\examples{
data_long <- data_wide \%>\% RadLab::wide_to_long(
  name_data = name_data,
  n_vignettes = 15,
  IDvar = "ID"
)
}
